completed fund buy transaction, working on price history transaction

+++ PortfolioManager2/PortfolioManager.Repository/Entities/InvestmentMap.cs
@@ -11,6 +11,6 @@ namespace PortfolioManager.Repository.Entities
         public int AccountId { get; set; }
         public int InvestmentId { get; set; }
         public decimal Quantity { get; set; }        
-        public int Valuation { get; set; }        
+        public decimal Valuation { get; set; }        
     }
 }
